Changed model logic again and further adjusted optimization service with loading previous routes, added cron jobs for clearing those routes.

+++ e_container/models/device.py
@@ -24,7 +24,7 @@ class DeviceModel(BaseModel):
     id = models.AutoField(primary_key=True)
     group = models.ForeignKey(DeviceGroupModel, null=True, blank=True)
     status = models.IntegerField(choices=DEVICE_STATUSES, default=NOT_INSTALLED)
-    max_capacity = models.PositiveIntegerField(help_text='Max capacity of the container it is installed on')
+    max_capacity = models.FloatField(help_text='Max capacity of the container it is installed on')
     type = models.IntegerField(choices=DEVICE_TYPES, default=STANDARD)
 
     def save(self, *args, **kwargs):
