+++ e_container/models/location.py
@@ -5,8 +5,8 @@ from e_container.models import MunicipalityModel
 
 class LocationModel(BaseModel):
     id = models.AutoField(primary_key=True)
-    latitude = models.DecimalField(max_digits=8, decimal_places=6)
-    longitude = models.DecimalField(max_digits=9, decimal_places=6)
+    latitude = models.FloatField()
+    longitude = models.FloatField()
     street = models.CharField(max_length=225)
     street_number = models.CharField(max_length=4, null=True, blank=True,
                                      help_text='Used for the division of longer streets')
